+++ examples/flow-react-example/src/basics/exportNestedValues.js
@@ -13,3 +13,5 @@ export class TopLevelClass {
 }
 
 export const ValueStartingWithUpperCaseLetter = "ValueStartingWithUpperCaseLetter";
+
+export default 42;